Separate both parser proc_macros into their own files
